refactor: Move NetworkingException to the service package
