+++ network/src/main/java/uk/gov/android/network/client/headers/AuthorisationHeaderReader.kt
@@ -1,11 +1,11 @@
 package uk.gov.android.network.client.headers
 
-import uk.gov.android.network.api.v2.ConfigurationException
-import uk.gov.android.network.api.v2.ServiceException
 import uk.gov.android.network.auth.AuthenticationProvider
 import uk.gov.android.network.auth.AuthenticationResponse
 import uk.gov.android.network.client.config.RequestConfig
 import uk.gov.android.network.http.Header
+import uk.gov.android.network.service.ConfigurationException
+import uk.gov.android.network.service.ServiceException
 import uk.gov.android.network.util.NetworkingResult
 
 private const val AUTH_HEADER_KEY = "Authorization"
