+++ network/src/test/java/uk/gov/android/network/client/headers/AuthorisationHeaderReaderTest.kt
@@ -4,12 +4,12 @@ import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertInstanceOf
 import org.junit.jupiter.api.Test
-import uk.gov.android.network.api.v2.ConfigurationException
-import uk.gov.android.network.api.v2.ServiceException
 import uk.gov.android.network.auth.TestAuthenticationProvider
 import uk.gov.android.network.auth.authenticationFailure
 import uk.gov.android.network.auth.authenticationSuccess
 import uk.gov.android.network.client.config.RequestConfig
+import uk.gov.android.network.service.ConfigurationException
+import uk.gov.android.network.service.ServiceException
 import uk.gov.android.network.util.expectFailure
 import uk.gov.android.network.util.expectSuccess
