+++ network/src/testFixtures/java/uk/gov/android/network/api/v2/ApiResponseAssertions.kt
@@ -2,9 +2,9 @@ package uk.gov.android.network.api.v2
 
 import org.junit.jupiter.api.Assertions.assertInstanceOf
 
-fun ApiResponse<*>.expectFailure(): ApiResponse.Failure = assertInstanceOf(ApiResponse.Failure::class.java, this)
+fun ApiResponse<*, *>.expectFailure(): ApiResponse.Failure<*> = assertInstanceOf(ApiResponse.Failure::class.java, this)
 
-fun <T> ApiResponse<T>.expectSuccess(): ApiResponse.Success<T> {
+fun <T> ApiResponse<T, *>.expectSuccess(): ApiResponse.Success<T> {
     assertInstanceOf(ApiResponse.Success::class.java, this)
     return this as ApiResponse.Success<T>
 }
